move misc Item classes to misc.h
